fix: comments of PR

+++ libs/engine/includes/Components/Animation.hpp
@@ -28,8 +28,8 @@ public:
      * @param offset The number of sprites in the animation.
      */
     Animation(
-        float textureWidth, float textureHeight, float entityWidth,
-        float entityHeight, float offset, size_t duration
+        size_t textureWidth, size_t textureHeight, size_t entityWidth,
+        size_t entityHeight, size_t offset, size_t duration
     )
         : _textureWidth(textureWidth)
         , _textureHeight(textureHeight)
@@ -45,12 +45,12 @@ public:
      */
     Animation() = default;
 
-    float _textureWidth = 0;
-    float _textureHeight = 0;
-    float _entityWidth = 0;
-    float _entityHeight = 0;
-    float _offset = 1; ///< The number of sprites in the animation.
-    float _currentOffset = 0; ///< The current index of the animation.
+    size_t _textureWidth = 0;
+    size_t _textureHeight = 0;
+    size_t _entityWidth = 0;
+    size_t _entityHeight = 0;
+    size_t _offset = 1; ///< The number of sprites in the animation.
+    size_t _currentOffset = 0; ///< The current index of the animation.
     size_t _duration = 0;
 };
 }
